Separate player listener wrapper from PlayerHolder
