+++ app/src/main/java/org/schabi/newpipe/info_list/dialog/InfoItemDialog.java
@@ -23,7 +23,7 @@ import org.schabi.newpipe.error.UserAction;
 import org.schabi.newpipe.extractor.InfoItem;
 import org.schabi.newpipe.extractor.stream.StreamInfoItem;
 import org.schabi.newpipe.extractor.stream.StreamType;
-import org.schabi.newpipe.player.helper.PlayerHolder;
+import org.schabi.newpipe.player.bind.PlayerHolder;
 import org.schabi.newpipe.util.StreamTypeUtil;
 import org.schabi.newpipe.util.external_communication.KoreUtils;
